requirements.txt
python-telegram-bot[job-queue]==20.0
yfinance==0.2.18
pandas==2.1.4
beautifulsoup4==4.12.3
requests==2.31.0
transformers==4.36.2
torch==2.2.1
flask==2.3.3
werkzeug==3.0.1
pytz==2024.1
numpy==1.26.4
gunicorn==21.2.0
protobuf==3.20.3
sentencepiece==0.2.0